requirements.txt
celery==5.3.6
Django==5.0.1
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
django-crontab==0.7.1
django-phonenumber-field==7.2.0
gunicorn==20.1.0
django-filter==23.4
django-choices==2.0.0
pre-commit==3.5.0
phonenumbers==8.13.26
requests==2.31.0
boto3==1.34.36
botocore==1.34.50
inflect==7.0.0

psycopg2-binary==2.9.9

